refactor: Refatora meu service e controller para tratar as exceções ao usuario

+++ src/main/java/br/com/mmaverse/service/ContenderService.java
@@ -2,6 +2,8 @@ package br.com.mmaverse.service;
 
 import br.com.mmaverse.entity.Contender;
 import br.com.mmaverse.entity.Gym;
+import br.com.mmaverse.exception.EntityNotFoundException;
+import br.com.mmaverse.exception.InvalidCpfException;
 import br.com.mmaverse.repository.ContenderRepository;
 import org.springframework.stereotype.Service;
 
@@ -9,7 +11,6 @@ import java.time.LocalDate;
 import java.time.Period;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ContenderService {
@@ -27,41 +28,43 @@ public class ContenderService {
     }
 
     public Contender save(Contender contender) {
+        if (repository.existsByCpf(contender.getCpf())) {
+            throw new InvalidCpfException("CPF already exists");
+        }
         contender.setAge(getAge(contender.getBirthDate()));
         contender.setGyms(findGym(contender.getGyms()));
         return repository.save(contender);
     }
 
-    public Optional<Contender> findById(Long id) {
-        return repository.findById(id);
+    public Contender findById(Long id) {
+        return repository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Contender not found with id: " + id));
     }
 
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new EntityNotFoundException("Contender not found with id: " + id);
+        }
         repository.deleteById(id);
     }
 
-    public Optional<Contender> update(Long id, Contender updateContender) {
-        Optional<Contender> optContender = repository.findById(id);
-        if (optContender.isPresent()) {
-            List<Gym> gyms = findGym(updateContender.getGyms());
+    public Contender update(Long id, Contender updateContender) {
+        Contender contender = findById(id);
+        List<Gym> gyms = findGym(updateContender.getGyms());
 
-            Contender contender = optContender.get();
-            contender.setName(updateContender.getName());
-            contender.setNickname(updateContender.getNickname());
-            contender.setCpf(updateContender.getCpf());
-            contender.setBirthDate(updateContender.getBirthDate());
-            contender.setAge(getAge(updateContender.getBirthDate()));
-            contender.setWeightClass(updateContender.getWeightClass());
-            contender.setHeight(updateContender.getHeight());
-            contender.setReach(updateContender.getReach());
-            contender.setWin(updateContender.getWin());
-            contender.setLose(updateContender.getLose());
-            contender.setGyms(gyms);
+        contender.setName(updateContender.getName());
+        contender.setNickname(updateContender.getNickname());
+        contender.setCpf(updateContender.getCpf());
+        contender.setBirthDate(updateContender.getBirthDate());
+        contender.setAge(getAge(updateContender.getBirthDate()));
+        contender.setWeightClass(updateContender.getWeightClass());
+        contender.setHeight(updateContender.getHeight());
+        contender.setReach(updateContender.getReach());
+        contender.setWin(updateContender.getWin());
+        contender.setLose(updateContender.getLose());
+        contender.setGyms(gyms);
 
-            repository.save(contender);
-            return Optional.of(contender);
-        }
-        return Optional.empty();
+        return repository.save(contender);
     }
 
     public Integer getAge(LocalDate dateOfBirth) {
@@ -70,8 +73,7 @@ public class ContenderService {
 
     public List<Gym> findGym(List<Gym> gyms) {
         List<Gym> gymsFound = new ArrayList<>();
-        gyms.forEach(gym -> gymService.findById(gym.getId())
-                .ifPresent(gymsFound::add));
+        gyms.forEach(gym -> gymsFound.add(gymService.findById(gym.getId())));
         return gymsFound;
     }
 }
